Modified linked lists into hash maps

+++ src/track/trackplaytimers.h
@@ -21,7 +21,7 @@ namespace TrackTimers {
     public:
       RegularTimer() = default;
       virtual ~RegularTimer() = default;
-      virtual void start(int msec) = 0;
+      virtual void start(double msec) = 0;
       virtual bool isActive() const = 0;
     public slots:
       virtual void stop() = 0;
@@ -34,7 +34,7 @@ namespace TrackTimers {
     public:
       GUITickTimer();
       ~GUITickTimer() override = default;
-      void start(int msec) override;
+      void start(double msec) override;
       bool isActive() const override;
       void stop() override;
     private:
